🐛 fix(TokenManagement.py): change return type of price_to_sqrt_price_x_96 from int to float The return type of the price_to_sqrt_price_x_96 method has been changed from int to float to ensure that the method returns a more accurate value. This change is necessary because the method is used to convert a price to a square root price, which requires a high degree of accuracy.

+++ package/uniswap_math/TokenManagement.py
@@ -56,7 +56,7 @@ class TokenManager:
         p = price / self.Q96
         return math.floor(math.log(p, base))
 
-    def price_to_sqrt_price_x_96(self, price: float) -> int:
+    def price_to_sqrt_price_x_96(self, price: float) -> float:
         """Converts a price to a sqrt price, useable by Uniswap V3
 
         Args:
@@ -65,7 +65,7 @@ class TokenManager:
         Returns:
             int: The converted sqrt price value
         """
-        return int(math.sqrt((self.token0 / self.token1) / price) * self.Q96)
+        return math.sqrt((self.token0 / self.token1) / price) * self.Q96
 
     def sqrt_price_x_96_to_price(self, sqrt_price_x_96: int) -> float:
         """Converts a sqrt price to a price, useable by Uniswap V3
